load sponsors via builder
